refactor: InterestTagApi 분리 및 uri 변경
#249
